Crate dungeon

Dungeon creation. Creates 2d cave-like dungeons using cellular automata (among other techniques).

Examples

use dungeon::{Dungeon, DungeonType};
use std::{
    convert::TryInto,
    num::NonZeroU16
};

let _ = Dungeon::new(
    150.try_into().unwrap(),
    100.try_into().unwrap(),
    NonZeroU16::new(10).unwrap(),
    DungeonType::Cave,
    false,
);

Re-exports

pub use crate::dungeon_tile::DungeonTile;

pub use point_index::PointIndex;

Modules

border
A border of points around a cave in a floor.

connection_path
A connection between two borders with a path drawn between them.

dungeon_tile
The various things a tile can be in a dungeon floor.

floor_builder
Builder struct for a Floor.

point
A point somewhere in a Floor.

point_index
A 1-dimensional type representing a 2-dimensional grid, indexable by a Point.

Structs

Column
Wrapper type around a BoundedInt that represents the column position of a Point. Note that the BoundedInt is bound on the maximum and minimum that a point can be, not the floor size itself.

Connection
A connection between two points on two different [Border]s.

Dungeon
A 2D dungeon containing multiple floors of various sizes.

Floor
A floor of a Dungeon.

FloorBuilder
Builder struct for a Floor.

FloorId
A unique, opaque ID assigned to each floor upon creation.

Point
A point somewhere in a Floor.

Row
Wrapper type around a BoundedInt that represents the row position of a Point. Note that the BoundedInt is bound on the maximum and minimum that a point can be, not the floor size itself.

Enums

DungeonType
The different types of dungeon a Dungeon can be.

Functions

distance
Returns the distance between two Points.