Module dungeon::floor_builder

Builder struct for a Floor.

See the type-level documentation for more information.

Modules

floor_builder_state

All the valid states a floor builder may be in. each module contains it’s own state, and the impls found in each module represent the valid transitions for each state.

to_block_character

Represents a type that can be ‘pretty-printed’ using ascii characters.

Structs

FloorBuilder

Builder struct for a Floor.

Constants

MAX_FLOOR_SIZE

The maximum dimensions a Floor can have.

MIN_FLOOR_SIZE

The minimum dimensions a Floor can have.

RANDOM_FILL_WALL_PERCENT_CHANCE

The percent chance of a wall being placed during the initial noise generation.