Module dungeon::floor_builder::floor_builder_state

All the valid states a floor builder may be in. each module contains it’s own state, and the impls found in each module represent the valid transitions for each state.

Modules

blank

A blank FloorBuilder, with everything set to their default.

drawable

A FloorBuilder that has state to write to it’s internal map.

filled

A resting state for the FloorBuilder. It is technically ‘completable’ in this state.

has_borders

A FloorBuilder that has the borders around all of the caves in it’s map.

has_connections

FloorBuilder state containing connections between the caves.

has_secret_connections

State that marks the FloorBuilder as having the secret paths drawn between the remaining caves that weren’t connected.

new

Original state of the FloorBuilder. Entry point to the state machine.

random_filled

First “real” state of the builder, where the noise map and tiles map have both been randomly filled.

smoothed

State that represents a FloorBuilder that has has the cellular automata run on it, smoothing out the edges of the caves.

Traits

FloorBuilderState

Marker trait for possible states for a floor builder to be in, along with their associated data.

Smoothable

Marker trait for floor builder states that can be FloorBuilder::smoothened.