Trait dungeon::floor_builder::floor_builder_state::FloorBuilderState

pub trait FloorBuilderState: Debug {
    const TYPE_NAME: &'static str;
}

Marker trait for possible states for a floor builder to be in, along with their associated data.

Associated Constants

const TYPE_NAME: &'static str

The name of the type. Used to allow for reflection (of sorts) when inspecting the floor builder state machine.

Loading content...