Struct dungeon::floor_builder::FloorBuilder

pub struct FloorBuilder<S: FloorBuilderState> {
    pub(crate) width: BoundedInt<MIN_FLOOR_SIZE, MAX_FLOOR_SIZE>,
    pub(crate) height: BoundedInt<MIN_FLOOR_SIZE, MAX_FLOOR_SIZE>,
    pub(crate) map: Vec<DungeonTile>,
    pub(crate) noise_map: Vec<u16>,
    extra: S,
    frames: Option<Vec<Frame<'static>>>,
    id: FloorId,
}

Builder struct for a Floor.

See http://roguebasin.roguelikedevelopment.org/index.php?title=Cellular_Automata_Method_for_Generating_Random_Cave-Like_Levels

Fields
width: BoundedInt<MIN_FLOOR_SIZE, MAX_FLOOR_SIZE>

The width of the floor.

height: BoundedInt<MIN_FLOOR_SIZE, MAX_FLOOR_SIZE>

The height of the floor.

map: Vec<DungeonTile>

The map of tiles.

noise_map: Vec<u16>

The map containing the noise for the floor builder, used in various places during the build process.

extra: S

The current state of the floor builder. May or may not contain extra data to be used at that stage of generation.

frames: Option<Vec<Frame<'static>>>

The frames of the floor builder as it is being built. Defaults to None unless specified to output to a gif.

id: FloorId

A unique, opaque ID assigned to the floor builder upon creation.

Implementations

impl FloorBuilder<Blank>

pub(in floor_builder) fn random_fill(
self
) -> FloorBuilder<RandomFilled>

TODO: Split this function into two parts, random_fill and trace_entrance_exit (or something along those lines)

impl FloorBuilder<Drawable>

pub(in floor_builder) fn draw(
self,
draw_with: fn(_: bool, _: bool, _: Point) -> DungeonTile
) -> FloorBuilder<Filled>

Draws the current state of the FloorBuilder with the provided function.

draw_with takes 3 parameters: The first is true if the point is the first in the path, the second is true if the point is the last in the path, and the third argument is the point itself. The function is expected to return a DungeonTile that will be placed at the provided point.

Note that the first and second arguments are not mutually exclisive.

impl FloorBuilder<Filled>

pub(in floor_builder) fn finish(
self
) -> Floor

Finishes the builder, returning the completed Floor and writing the gif out to out/frame_{id}.gif.

impl FloorBuilder<HasBorders>

pub(in floor_builder) fn build_connections(
self,
iterations: BuildConnectionIterations
) -> FloorBuilder<HasConnections>

Builds bridges between the disjointed caves and the closest cave border point not in the border of the first cave.

impl FloorBuilder<HasConnections>

pub(in floor_builder) fn trace_connection_paths(
self,
wide: bool,
use_noise_map: bool
) -> FloorBuilder<Drawable>

Takes the connections from FloorBuilder<HasBorders>::build_connections and traces paths between them, leaving the paths in the to_draw state of FloorBuilder<Drawable>.

impl FloorBuilder<HasSecretPassages>

pub(in floor_builder) fn place_treasure_chests(
self
) -> FloorBuilder<Filled>

Places the treasure chests on the map.

impl FloorBuilder<New>

pub(crate) fn create(
id: FloorId,
width: BoundedInt<MIN_FLOOR_SIZE, MAX_FLOOR_SIZE>,
height: BoundedInt<MIN_FLOOR_SIZE, MAX_FLOOR_SIZE>,
gif_output: bool
) -> Floor

Creates a new floor builder with the provided values.

impl FloorBuilder<RandomFilled>

pub(in floor_builder) fn trace_original_path(
self
) -> FloorBuilder<Filled>

Traces the original path through the map from the entrance to the exit.

impl FloorBuilder<Smoothed>

pub(in floor_builder) fn get_cave_borders(
self
) -> FloorBuilder<HasBorders>

Finds the borders around all of the caves in the FloorBuilder, leaving them in the HasBorders state.

pub(in floor_builder) fn check_for_secret_passages(
self
) -> FloorBuilder<HasSecretPassages>

Adds secret passages to the map between all of the remaining disjointed caves.

impl<S: Smoothable> FloorBuilder<S>

fn smoothen(
self,
repeat: usize,
create_new_walls: fn(_: usize) -> bool
) -> FloorBuilder<Smoothed>

Smooths out the map using cellular automata.

impl<S: FloorBuilderState> FloorBuilder<S>

fn inspect(self) -> Self

Dumps the current state of the builder to stdout. for use in debugging only

fn frame_from_current_state(&mut self, delay: u16)

Creates a [gif::Frame] from the current state of the floor builder, storing it in FloorBuilder::frames.

Note that this is essentially a noop if self.frames is None (i.e., gif output has not been enabled).

fn blank(
id: FloorId,
width: BoundedInt<MIN_FLOOR_SIZE, MAX_FLOOR_SIZE>,
height: BoundedInt<MIN_FLOOR_SIZE, MAX_FLOOR_SIZE>,
gif_output: bool
) -> FloorBuilder<Blank>

Creates a blank FloorBuilder, with all the values set to their defaults.

fn place_wall_logic(&self, point: Point, create_new_walls: bool) -> DungeonTile

will only return wall or empty

pub fn get_adjacent_walls(
&self,
point: Point,
distance_x: i32,
distance_y: i32
) -> usize

Returns how many walls there are within the rectangle bounded by distance_x left and right of the point and distance_y above and below the point.

pub fn is_wall(&self, point: Point) -> bool

Considers out-of-bounds a wall

fn is_out_of_bounds(&self, point: Point) -> bool

Considers the 1-wide border around the edge of the map to be out of bounds.

fn get_legal_neighbors_with_diagonals(
&self,
point: Point
) -> impl Iterator<Item = Point> + '_

Gets the 8 neighbours around the specified point that aren’t out of bounds.

x x x
x p x
x x x

fn get_legal_neighbors(&self, point: Point) -> impl Iterator<Item = Point> + '_

Gets the 4 neighbours around the specified point that aren’t out of bounds.

o x o
x p x
o x o

fn get_legal_neighbors_down_and_right(
&self,
point: Point
) -> impl Iterator<Item = Point> + '_

Gets the 2 neighbours around the specified point that aren’t out of bounds, below it and to the right.

o o o
o p x
o x 0

pub(crate) fn _pretty(
&self,
extra_points: &[Point],
extra_points2: &[Point]
) -> String

Pretty-prints the map in it’s current state. Used for debugging and tests.